+++ lib/src/bones_api_entity.dart
@@ -28,6 +28,7 @@ import 'bones_api_mixin.dart';
 import 'bones_api_platform.dart';
 import 'bones_api_types.dart';
 import 'bones_api_utils.dart';
+import 'bones_api_utils_call.dart';
 import 'bones_api_utils_collections.dart';
 import 'bones_api_utils_instance_tracker.dart';
 import 'bones_api_utils_json.dart';