+++ lib/src/bones_api_entity_db_object_gcs.dart
@@ -274,7 +274,8 @@ class DBObjectGCSAdapter extends DBObjectAdapter<DBObjectGCSAdapterContext> {
 
   @override
   TableScheme? getTableSchemeImpl(
-      String table, TableRelationshipReference? relationship) {
+      String table, TableRelationshipReference? relationship,
+      {Object? contextID}) {
     _log.info('getTableSchemeImpl> $table ; relationship: $relationship');
 
     var tableScheme = tablesSchemes[table];